Changed repository name and building GeneriKit will be known as a hub for generic drivers. It is better to take them apart and program each one by a separate repository, then add a GeneriKit installer.
There are, still, a lot of stuff to do.
This repository will only be used to build generic PlayStation devices or fixes.

After 10+ or 100+ devices/fixes, it may be relevant to gather all up and push to the main kernel.
