+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "vendor/openvr"]
-	path = vendor/openvr
-	url = https://github.com/KanuX-14/openvr.git
+[submodule "vendor/bluez"]
+	path = vendor/bluez
+	url = https://github.com/KanuX-14/bluez.git
